+++ CHANGES.md
@@ -7,6 +7,9 @@ Unreleased
 - Allow overriding the `ocaml` binary with findlib configuration (#7648,
   @rgrinberg)
 
+- merlin: ignore instrumentation settings for preprocessing. (#7606, fixes
+  #7465, @Alizter)
+
 - When a rule's action is interrupted, delete any leftover directory targets.
   This is consistent with how we treat file targets. (#7564, @rgrinberg)